% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapp_hg19}
\alias{mapp_hg19}
\title{GRanges with mappability scores for hg19}
\format{A GRanges object with 21591667 ranges and mappability scores}
\usage{
mapp_hg19
}
\description{
GRanges of mappability track for 100-mers on the GRCh37/hg19 human reference
        genome from ENCODE.
}
\keyword{datasets}
