% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapp_hg38}
\alias{mapp_hg38}
\title{GRanges with mappability scores for hg38}
\format{A GRanges object with 21584930 ranges and mappability scores}
\usage{
mapp_hg38
}
\description{
Use liftOver utility to convert hg19 coordinates to hg38
}
\keyword{datasets}
